+++ Install-Croupier-Droid.py
@@ -1,16 +1,18 @@
-import ssl
-from os import path, getcwd, chdir, walk, mkdir, remove, system, listdir
-from urllib import request
-from zipfile import ZipFile
+# Import select modules from OS package
+from os import path, getcwd, chdir, mkdir, walk, remove, system, listdir
+# Import move/rename module and recursive delete module
 from shutil import move, rmtree
-from sys import path as syspath
-from importlib import import_module
-from re import findall, fullmatch     
 
+# Purpose: Self-update installer/updater and download source code
+# Parameters: None
+# Returns: None
 def UpdateSelf():
+    # import ssl package
+    import ssl
     # Bypass SSL certificate errors
     ssl._create_default_https_context = ssl._create_unverified_context
 
+    # Capture current working directory
     originalPath = getcwd()
 
     print("Downloading updates ...")
@@ -41,14 +43,19 @@ def UpdateSelf():
 
     # Navigate to source directory
     chdir(sourcePath)
-
+    
+    # Import downloader module and internet error module
+    from urllib.request import urlretrieve
+    from urllib.error import URLError
     # Download repo as ZIP file
     try:
-        request.urlretrieve("https://github.com/chandler-stevens/Croupier-Droid/archive/master.zip", "./Croupier-Droid-master.zip")
-    except urllib.error.URLError:
+        urlretrieve("https://github.com/chandler-stevens/Croupier-Droid/archive/master.zip", "./Croupier-Droid-master.zip")
+    except URLError:
         input("ERROR! No Internet connection!")
         exit()
 
+    # Import ZIP extractor module
+    from zipfile import ZipFile
     # Extract ZIP file
     repo = ZipFile(sourcePath + "Croupier-Droid-master.zip", "r")
     repo.extractall(sourcePath)
@@ -58,20 +65,29 @@ def UpdateSelf():
     move(sourcePath + "Croupier-Droid-master/Install-Croupier-Droid.py", installPath + "Temp.py")
     move(sourcePath + "Croupier-Droid-master/__init__.py", installPath)
 
-    # Load new Updater
+    # Import path marker module
+    from sys import path as syspath
+    # Mark installation path as import location
     syspath.append(installPath)
+    
+    # Import importer module
+    from importlib import import_module
+    # Import updater initialization module
     module = import_module("__init__")
 
     # Install updates
     module.UpdateSource(installPath, sourcePath, originalPath)
 
 
+# Purpose: Move files and overwrite any existing conflicts
+# Parameters: Source path (string), destination path (string)
+# Returns: None
 def Update(source, destination):
     # Move/Overwrite all version files
     for src_dir, dirs, files in walk(source):
         # Determine destination path
         dst_dir = src_dir.replace(source, destination)
-        # Create destination directory, if necessary
+        # Create destination directory if necessary
         if not path.exists(dst_dir):
             mkdir(dst_dir)
         # For each new version file
@@ -85,7 +101,13 @@ def Update(source, destination):
             # Move file
             move(src_file, dst_dir)
 
+# Purpose: Updates source code to latest stable version(s)
+# Parameters: Installation path (string), installed source path (string), this file's path (string) 
+# Returns: None
 def UpdateSource(installPath, sourcePath, originalPath):
+    # Import regex package
+    from re import findall, fullmatch
+    
     # Cleanup launchers
     for fileName in listdir(installPath):
         filePath = path.join(installPath, fileName)
@@ -135,7 +157,7 @@ def UpdateSource(installPath, sourcePath, originalPath):
     if path.isfile(sourcePath + "Croupier-Droid-master.zip"):
         remove(sourcePath + "Croupier-Droid-master.zip")
 
-    # Hide source code
+    # Hide source code in Windows
     if sourcePath[0] == "C":
         from ctypes import windll
         windll.kernel32.SetFileAttributesW(sourcePath, 2)
@@ -144,7 +166,9 @@ def UpdateSource(installPath, sourcePath, originalPath):
 
 
 try:
+    # Self-update updater if not done so already
     if not path.isfile("C:/CroupierDroid/Temp.py") and not path.isfile("/storage/emulated/0/qpython/CroupierDroid/Temp.py"):
         UpdateSelf()
+# Capture any installation errors
 except Exception as error:
         input("ERROR! " + str(error))